Win: fix OpenProcess not recognizing when proc is gone.
This is due to ExitCodeProcess needing
PROCESS_QUERY_INFORMATION access rights.

+++ psutil/__init__.py
@@ -1227,6 +1227,9 @@ class Process(object):
             self._send_signal(sig)
         else:  # pragma: no cover
             self._raise_if_pid_reused()
+            if sig != signal.SIGTERM and not self.is_running():
+                msg = "process no longer exists"
+                raise NoSuchProcess(self.pid, self._name, msg=msg)
             self._proc.send_signal(sig)
 
     def suspend(self):
